some searcher cleanup, spot selector for phrases containing nouns, index enricher for priors

+++ core/src/main/java/org/dbpedia/spotlight/lucene/search/BaseSearcher.java
@@ -68,7 +68,7 @@ public class BaseSearcher implements Closeable {
         LOG.debug("Done.");
     }
 
-    public long getNumberOfEntries() {
+    public int getNumberOfEntries() {
         return this.mReader.numDocs();    // can use maxDoc?
     }
